requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Markdown==3.6
python-dotenv==1.0.1
Werkzeug==3.0.3
